[Mosaic GPU] Introduce constraints in the equational layout inference.
This allows fixing a quirk in hint derivation, by ensuring that wrong choices
of hints can be backtracked from safely. The data flow of the graph imposes
relayout constraints such that, for a variable `v`

* `Relayout(producer(v), v)`, and
* `forall c : consumers(v), Relayout(v, c)`

must hold. We now derive these constraints automatically.

As a consequence of the change, `test_infer_splat_layout_for_vector_splat`
needed to modify to either allow propagating `splat`s up towards the `undef`
op (in which case, the `undef` op's output would always be assigned a `splat`
layout), or forbidding `splat` layouts for the `undef` op's output
systematically (in which case, the `undef` op's output would always be
assigned a `strided` layout). We decided to go with the latter, and eliminated
the splat case from the test.

PiperOrigin-RevId: 783595782